requirements.txt
accelerate==1.12.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
distro==1.9.0
durationpy==0.10
filelock==3.20.3
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
oauthlib==3.3.1
onnxruntime==1.24.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pillow==12.1.0
posthog==5.4.0
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.3
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.46
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.1.0
typer==0.24.0
typer-slim==0.21.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.3
zipp==3.23.0
